Moved getAccuracy over to predict.py as that is where the evaluation of model quality should be.
